third_party/blink/web_tests/platform/generic/fast/mediacapturefromelement/CanvasCaptureMediaStream-request-frame-events-expected.txt

Exercises the requestFrame() call on captureStream(0).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Got a stream from canvas.
PASS stream.getVideoTracks().length is 1
PASS track.readyState is "live"
PASS Plugged stream to video tag.
PASS Video play callback succeeded.
PASS Drawing to canvas.
PASS Calling requestFrame().
PASS Video canplaythrough callback succeeded.
PASS successfullyParsed is true

TEST COMPLETE

 
